pyproject.toml
[tool.poetry]
name = "clickclickmoney-ai"
version = "0.1.0"
description = "LangChain 기반 개인 재무 코치 챗봇"
authors = ["HyunGeunK <hyungeun0313@naver.com>"]
readme = "README.md"
packages = [{ include = "clickclickmoney_ai", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11,<3.15"

# --- Core App & Web Server ---
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"

# --- LangChain & AI ---
langchain = "^0.2.5"
langchain-community = "^0.2.5"
langchain-openai = "^0.1.8"
langgraph = "^0.1.1"
openai = "^1.35.3"
tiktoken = "^0.7.0"

# --- Database & Vector Store ---
sqlalchemy = "^2.0.31"
mariadb = "^1.1.12"      # MariaDB 직접 연결용 (선택)
pymysql = "^1.1.1"      # SQLAlchemy가 MariaDB에 연결할 때 사용하는 드라이버
faiss-cpu = "^1.8.0"     # RAG를 위한 벡터 DB

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"